requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.17.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
xlrd>=2.0.1

